requirements.txt
pyyaml
notify2
requests
dbus-python